🚚 Integrate tournament settings into management page
