+++ src/main/webui/src/components/pages/competition/ViewCreateCompetition.vue
@@ -5,7 +5,7 @@
 			<template #content>
 				<FormCompetition
 					ref="form"
-					:competition="competition"
+					:competition="CompetitionDefault"
 					:disabled="false"
 					@submit="submit"
 				/>
@@ -14,6 +14,7 @@
 				<div class="justify-content-end flex">
 					<Button
 						:label="t('general.create')"
+						severity="success"
 						@click="() => form !== null && form.onSubmit()"
 					/>
 				</div>
@@ -24,20 +25,11 @@
 
 <script lang="ts" setup>
 import { router } from "@/main"
-import { reactive, ref } from "vue"
+import { ref } from "vue"
 import { useRoute } from "vue-router"
 import FormCompetition from "@/components/pages/forms/FormCompetition.vue"
 import { useI18n } from "vue-i18n"
-import {
-	Competition,
-	CompetitionServer,
-	Mode,
-	NumberSets,
-	Progress,
-	Sex,
-	SignUp,
-	CompType,
-} from "@/interfaces/competition"
+import { CompetitionServer, CompetitionDefault } from "@/interfaces/competition"
 import { useToast } from "primevue/usetoast"
 import { useAddCompetition } from "@/backend/competition"
 import { Routes } from "@/routes"
@@ -47,31 +39,6 @@ const toast = useToast()
 
 const route = useRoute()
 const form = ref<InstanceType<typeof FormCompetition> | null>(null)
-
-const competition = reactive<Competition>({
-	name: "",
-	description: "",
-	tourType: CompType.KNOCKOUT,
-	mode: Mode.SINGLE,
-	signUp: SignUp.INDIVIDUAL,
-	numberSets: NumberSets.THREE,
-	playerA: {
-		sex: Sex.ANY,
-		hasMinAge: false,
-		minAge: null,
-		hasMaxAge: false,
-		maxAge: null,
-	},
-	playerB: {
-		different: false,
-		sex: Sex.ANY,
-		hasMinAge: false,
-		minAge: null,
-		hasMaxAge: false,
-		maxAge: null,
-	},
-	cProgress: Progress.TEAMS,
-})
 
 const { mutate } = useAddCompetition(route, t, toast, {
 	suc: () => {
