+++ src/main/webui/src/components/pages/forms/FormTournament.vue
@@ -1,112 +1,93 @@
 <template>
-	<div class="flex justify-content-center w-full">
-		<Card id="card">
-			<template #title>{{ t(props.header) }}</template>
-			<template #content>
-				<div class="formgrid grid">
-					<div class="field flex w-full justify-content-end">
-						<div class="col" style="flex-grow: 3">
-							<label for="name">{{ t("general.name") }}</label>
-							<InputText
-								id="name"
-								v-model="name"
-								type="text"
-								v-bind="nameAttrs"
-								maxlength="30"
-								class="w-full"
-								:class="{ 'p-invalid': errors.name }"
-								:disabled="disabled"
-							/>
-							<InlineMessage v-if="errors.name" class="mt-2"
-								>{{ t(errors.name || "") }}
-							</InlineMessage>
-						</div>
-						<div class="col flex flex-column min-w-max flex-grow-0">
-							<label for="visible" class="text-900">{{
-								t("TournamentSettings.visible")
-							}}</label>
-							<SelectButton
-								v-model="visible"
-								v-bind="visibleAttrs"
-								:unselectable="false"
-								:options="[
-									{ name: t('no'), value: false },
-									{ name: t('yes'), value: true },
-								]"
-								option-label="name"
-								option-value="value"
-								:disabled="disabled"
-							/>
-						</div>
-					</div>
-					<div class="field col-12">
-						<label for="description">{{ t("general.description") }}</label>
-						<Textarea
-							id="description"
-							v-model="description"
-							type="text"
-							rows="4"
-							maxlength="255"
-							class="w-full"
-							v-bind="descriptionAttrs"
-							:disabled="disabled"
-						></Textarea>
-					</div>
-					<div class="field col-6">
-						<label for="registration">{{
-							t("TournamentSettings.registration_phase")
-						}}</label>
-						<Calendar
-							id="registration"
-							v-model="registrationPhase"
-							v-bind="registrationPhaseAttrs"
-							selection-mode="range"
-							:manual-input="false"
-							class="w-full"
-							:date-format="t('date_format')"
-							show-time
-							show-icon
-							:class="{ 'p-invalid': errors.registration_phase }"
-							:disabled="disabled"
-						/>
-						<InlineMessage v-if="errors.registration_phase" class="mt-2"
-							>{{ t(errors.registration_phase || "") }}
-						</InlineMessage>
-					</div>
-					<div class="field col-6">
-						<label for="game_phase">{{
-							t("TournamentSettings.game_phase")
-						}}</label>
-						<Calendar
-							id="game_phase"
-							v-model="gamePhase"
-							v-bind="gamePhaseAttrs"
-							selection-mode="range"
-							:manual-input="false"
-							class="w-full"
-							:date-format="t('date_format')"
-							show-time
-							show-icon
-							:class="{ 'p-invalid': errors.game_phase }"
-							:disabled="disabled"
-						/>
-						<InlineMessage v-if="errors.game_phase" class="mt-2"
-							>{{ t(errors.game_phase || "") }}
-						</InlineMessage>
-					</div>
-				</div>
-				<div></div>
-			</template>
-			<template #footer>
-				<div class="justify-content-end flex">
-					<Button
-						:label="props.submitText"
-						:disabled="disabled"
-						@click="onSubmit"
-					></Button>
-				</div>
-			</template>
-		</Card>
+	<div class="formgrid grid">
+		<div class="field flex w-full justify-content-end">
+			<div class="col" style="flex-grow: 3">
+				<label for="name">{{ t("general.name") }}</label>
+				<InputText
+					id="name"
+					v-model="name"
+					type="text"
+					v-bind="nameAttrs"
+					maxlength="30"
+					class="w-full"
+					:class="{ 'p-invalid': errors.name }"
+					:disabled="disabled"
+				/>
+				<InlineMessage v-if="errors.name" class="mt-2"
+					>{{ t(errors.name || "") }}
+				</InlineMessage>
+			</div>
+			<div class="col flex flex-column min-w-max flex-grow-0">
+				<label for="visible" class="text-900">{{
+					t("TournamentSettings.visible")
+				}}</label>
+				<SelectButton
+					v-model="visible"
+					v-bind="visibleAttrs"
+					:unselectable="false"
+					:options="[
+						{ name: t('no'), value: false },
+						{ name: t('yes'), value: true },
+					]"
+					option-label="name"
+					option-value="value"
+					:disabled="disabled"
+				/>
+			</div>
+		</div>
+		<div class="field col-12">
+			<label for="description">{{ t("general.description") }}</label>
+			<Textarea
+				id="description"
+				v-model="description"
+				type="text"
+				rows="4"
+				maxlength="255"
+				class="w-full"
+				v-bind="descriptionAttrs"
+				:disabled="disabled"
+			></Textarea>
+		</div>
+		<div class="field col-6">
+			<label for="registration">{{
+				t("TournamentSettings.registration_phase")
+			}}</label>
+			<Calendar
+				id="registration"
+				v-model="registrationPhase"
+				v-bind="registrationPhaseAttrs"
+				selection-mode="range"
+				:manual-input="false"
+				class="w-full"
+				:date-format="t('date_format')"
+				show-time
+				show-icon
+				:class="{ 'p-invalid': errors.registration_phase }"
+				:disabled="disabled"
+			/>
+			<InlineMessage v-if="errors.registration_phase" class="mt-2"
+				>{{ t(errors.registration_phase || "") }}
+			</InlineMessage>
+		</div>
+		<div class="field col-6">
+			<label for="game_phase">{{ t("TournamentSettings.game_phase") }}</label>
+			<Calendar
+				id="game_phase"
+				v-model="gamePhase"
+				v-bind="gamePhaseAttrs"
+				selection-mode="range"
+				:manual-input="false"
+				class="w-full"
+				:date-format="t('date_format')"
+				show-time
+				show-icon
+				:class="{ 'p-invalid': errors.game_phase }"
+				:disabled="disabled"
+			/>
+			<InlineMessage v-if="errors.game_phase" class="mt-2"
+				>{{ t(errors.game_phase || "") }}
+			</InlineMessage>
+		</div>
 	</div>
 </template>
 
@@ -120,7 +101,6 @@ import {
 import { useI18n } from "vue-i18n"
 
 import { useForm } from "vee-validate"
-import Card from "primevue/card"
 
 import { array, boolean, date, mixed, object, setLocale, string } from "yup"
 import { toTypedSchema } from "@vee-validate/yup"
@@ -129,10 +109,8 @@ const { t } = useI18n()
 
 const props = withDefaults(
 	defineProps<{
-		submitText: string
 		disabled: boolean
 		tourDetails?: Tournament
-		header: string
 	}>(),
 	{
 		disabled: false,
@@ -212,16 +190,12 @@ const [registrationPhase, registrationPhaseAttrs] =
 	defineField("registration_phase")
 const [gamePhase, gamePhaseAttrs] = defineField("game_phase")
 
-const emit = defineEmits(["submit"])
-
 const onSubmit = handleSubmit((values) => {
 	values.id = data.id
 	emit("submit", tournamentFormClientToServer(<TournamentForm>values))
 })
+defineExpose({ onSubmit })
+const emit = defineEmits(["submit"])
 </script>
 
-<style scoped>
-#card {
-	width: min(90dvw, 50rem);
-}
-</style>
+<style scoped></style>
