+++ src/main/webui/src/components/pages/management/ViewManageCompetitions.vue
@@ -7,7 +7,7 @@
 			<template #header>
 				<ViewMenuSelector />
 				<ViewExecutionMenuSelector v-if="route.meta.executionMenu" />
-				<ViewCompSelector v-if="route.meta.compSelector" />
+				<ViewSelector v-if="route.meta.compSelector" />
 			</template>
 			<template #content>
 				<router-view />
@@ -17,7 +17,7 @@
 </template>
 
 <script setup lang="ts">
-import ViewCompSelector from "@/components/pages/management/tabs/ViewCompSelector.vue"
+import ViewSelector from "@/components/pages/management/tabs/ViewSelector.vue"
 import ViewMenuSelector from "@/components/pages/management/tabs/ViewMenuSelector.vue"
 import ViewExecutionMenuSelector from "@/components/pages/management/tabs/ViewExecutionMenuSelector.vue"
 import { inject, ref, watch } from "vue"
