+++ src/main/webui/src/components/pages/settings/SettingsCourts.vue
@@ -1,7 +1,11 @@
 <template>
 	<DataTable :value="courts" striped-rows class="w-full mb-2" size="small">
 		<Column field="name" :header="t('general.name')"></Column>
-		<Column field="courtType" :header="t('court.type')"></Column>
+		<Column field="courtType" :header="t('court.type')">
+			<template #body="{ data }">
+				{{ data.courtType ? t("court.courtTypes." + data.courtType) : "" }}
+			</template>
+		</Column>
 	</DataTable>
 	<ViewCreateCourtSmall />
 </template>
